rename sources for pair style lebedeva to be consistent with style name
