+++ src/Purge.list
@@ -24,6 +24,9 @@ style_nstencil.h
 style_ntopo.h
 # other auto-generated files
 lmpinstalledpkgs.h
+# renamed on 7 January 2019
+pair_lebedeva.cpp
+pair_lebedeva.h
 # removed on 12 December 2018
 pair_reax.cpp
 pair_reax.h
